Backend(,Tests.End2End): add a config dir for RunMode.Testing
The wallet has a lot of global variables which makes it hard to
encapsulate separate instances of geewallet for the sake of testing and
to keep tests from touching the real user wallet and config files.
Ideally this should be fixed by having some kind of `Geewallet` type
as a main entry point to the backend which contains all of the backend's
state. In the meantime though, the run-mode setting introduced in this
commit allows us to run tests without touching the main wallet.

GWallet.Backend.Config now has a RunMode type defined as

    type RunMode =
        | Normal
        | Testing of DirectoryInfo

The run mode must be set to either `Normal` or `Testing`. Tests will set
it to `Testing` by calling `Config.SetRunModeTesting`.