Fix frontend running on regtest
When running on regtest we need to use the locally-running electrum
server with a consistency config of 1.

Both these things were being overridden. The server list was being
updated from the frontend using a server list downloaded off of github
and the consistency config could sometimes be reset to 2. This is now
fixed by only downloading the server list when not on regtest and always
using a consistency config of 1 on regtest.