WIP - mine full blocks faster
To set bitcoind's fee rate you need to mine full blocks. This is both so
that you can push out transactions whose fees are too low and so that
bitcoind can infer the fee rate from the transactions in the blocks.

I was doing this by mining full blocks filled with ordinary
transactions, but this turned out to be too slow to be practical. This
WIP commit now mines blocks with giant (non-standard) transactions that
can be generated quickly.

The code is in a state where I've validated the proof-of-concept, I just
need to put everything back in order.