WIP - fixing fee setting (again)
Fee setting broke. This time because bitcoind would run out of spare
utxos to spend when generating the blocking-filling transactions. I'm
changing the block-filling code so that it always makes sure it has a
reserve of utxos by adding a transaction that splits the large coinbase
utxos that it gets from mining into hundreds of smaller ones. It's
currently not working because bitcoind will prefer to use and collect
lots of small utxos rather than break the big ones, so I'll have to
generate the splitting transaction more manually rather than just using
bitcoind's sendtomany function.