WIP[CPFP] - fee rates now start going down when trying to mine them upwards :/
The fee rate can be set, but later on in the test when trying to
increase it the fee rate goes up part-way then sinks back down again. It
starts crawling up again slowly but by then the test gives up because
it's taking too long.

I'm suspect this fee-drop is because bitcoind is using change txouts
from previous (low fee-rate) unconfirmed transactions in order to build
the newer (high fee-rate) transactions that I'm trying to fill blocks
with. Which means it still has to mine the low-fee-rate transactions
for the high-fee-rate transactions to be spendable.

This is already taken a silly amount of time for something that seemed
like it would be easy. I'm going to put it on the backburner now and
use the ElectrumClient fee-rate mocking that we had before. I can use
the block-filling code to stop cheap transactions getting mined in the
CPFP test. So I'll do that but not bother to try and set bitcoind's
reported fee rate.