Backend.Tests.End2End: finish regtest geewallet->lnd channel open
This test uses the regtest network to test that geewallet can
open a channel with lnd.

To run this test locally you just need to have bitcoind,
electrumx_server (the spesmilo version, which didn't drop bitcoin
support) and lnd in your `$PATH`. It shouldn't interfere with any local
config for these programs.